django-server/requirements.txt
# django-server/requirements.txt
Django==5.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
bcrypt==4.1.1
